Update: start-with-project-stub.ru.md
I added a link to the original article which was written in Russian to this page.  The next step is to actually convert that article into Markdown and publish that article here.

Does anyone have any input on this?  Should we for some reason *not* supply the Russian version here?